Fix the certificate resource to properly work
